requirements.txt
pyserial
pyodbc
logging
pymcprotocol
requests
pyyaml
sockets
xlsxwriter
websocket
paho-mqtt
pythonnet
numpy
cryptography